+++ sdk/src/assertions/mod.rs
@@ -14,8 +14,10 @@
 //! Assertion helpers to build, validate, and parse assertions.
 
 mod actions;
+#[allow(unused)]
+pub(crate) use actions::source_type;
 pub(crate) use actions::V2_DEPRECATED_ACTIONS;
-pub use actions::{c2pa_action, source_type, Action, ActionTemplate, Actions, SoftwareAgent};
+pub use actions::{c2pa_action, Action, ActionTemplate, Actions, SoftwareAgent};
 
 mod asset_reference;
 pub use asset_reference::AssetReference;
